Refactor: type을 전역으로 선언
