+++ index.html
@@ -84,6 +84,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="/src/ts/index.ts"></script>
+    <script type="module" src="/src/ts/App.ts"></script>
   </body>
 </html>
