+++ src/ts/Store.ts
@@ -1,20 +1,3 @@
-type categoryIndex =
-  | "espresso"
-  | "frappuccino"
-  | "blended"
-  | "teavana"
-  | "dessert";
-
-interface menuItemInfo {
-  category: string;
-  menuName: string;
-  status: string;
-}
-
-type menuItems = {
-  [k in categoryIndex]?: menuItemInfo;
-};
-
 class Store {
   setLocalStorage = (category: string, menuItems: menuItems) => {
     return localStorage.setItem(category, JSON.stringify(menuItems));
@@ -24,4 +7,4 @@ class Store {
   };
 }
 
-export const store = new Store();
+export default new Store();
