+++ src/ts/menu.d.ts
@@ -1,0 +1,16 @@
+interface MenuItemInfo {
+  menuName: string;
+  category: string;
+  status: string;
+}
+
+type categoryIndex =
+  | "espresso"
+  | "frappuccino"
+  | "blended"
+  | "teavana"
+  | "dessert";
+
+type menuItems = {
+  [k in categoryIndex]: object[];
+};
